+++ packages/math-input/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.11.1](https://github.com/pie-framework/pie-lib/compare/@pie-lib/math-input@6.11.0...@pie-lib/math-input@6.11.1) (2023-04-19)
+
+
+### Bug Fixes
+
+* **eslint:** fix eslint problems-PD-2732 ([9eb9976](https://github.com/pie-framework/pie-lib/commit/9eb9976749753da86e1057a07bd2cfc65bf64ae4))
+* **math-input:** add required import for MathQuill-PD-2732 ([532d786](https://github.com/pie-framework/pie-lib/commit/532d7866fa48a5f9ac68f6d26f31b0e28ac54acf))
+
+
+
+
+
 # [6.11.0](https://github.com/pie-framework/pie-lib/compare/@pie-lib/math-input@6.10.13...@pie-lib/math-input@6.11.0) (2023-04-07)